+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Default table block receives govuk styling
+
 ## [0.1.1] - 2020-05-26
 
 ### Changed